+++ src/elements/User.php
@@ -634,11 +634,6 @@ class User extends Element implements IdentityInterface
     private $_groups;
 
     /**
-     * @var array|null The user’s preferences
-     */
-    private $_preferences;
-
-    /**
      * @inheritdoc
      */
     public function init()
@@ -1314,11 +1309,7 @@ class User extends Element implements IdentityInterface
      */
     public function getPreferences(): array
     {
-        if ($this->_preferences === null) {
-            $this->_preferences = Craft::$app->getUsers()->getUserPreferences($this->id);
-        }
-
-        return $this->_preferences;
+        return Craft::$app->getUsers()->getUserPreferences($this->id);
     }
 
     /**
@@ -1378,12 +1369,11 @@ class User extends Element implements IdentityInterface
      *
      * @param array $preferences The new preferences
      * @return array The user’s new preferences.
+     * @deprecated in 3.7.27.
      */
     public function mergePreferences(array $preferences): array
     {
-        $this->_preferences = array_merge($this->getPreferences(), $preferences);
-
-        return $this->_preferences;
+        return $preferences + $this->getPreferences();
     }
 
     /**